school-proj/faceDetectionWithCamera/requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.24.3
mysql-connector-python==8.1.0
redis==5.0.0
persiantools==4.1.0
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0
boto3==1.28.85
python-dotenv==1.0.0
